requirements.txt
streamlit
audio_recorder_streamlit
python-dotenv
google-generativeai